+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `catch2TestExplorer.googletest.gmockVerbose` to the configs.
 - `catch2TestExplorer.googletest.treatGmockWarningAs` to the configs.
 
 ## [2.4.14] - 2019-06-13